backend/requirements.txt
Flask
flask-cors
openai
psycopg2-binary